server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
wsproto==1.2.0
